refactor: tidy up folder structure
