+++ packages/graphback-db-manage/src/index.ts
@@ -1,3 +1,3 @@
-export * from './DatabaseManager';
-export * from './database';
-export * from './providers';
+export { DatabaseManager } from './DatabaseManager';
+export { DatabaseInitializationStrategy, CreateDatabaseIfNotExists, UpdateDatabaseIfChanges, DropCreateDatabaseAlways, DropCreateDatabaseIfChanges, DatabaseContextProvider, DefaultDataContextProvider, SchemaProvider, InputModelProvider, DatabaseSchemaManager } from './database';
+export { KnexMigrationProvider } from './migrations';
