--- packages/graphback-db-manage/src/models/index.ts
@@ -1,1 +1,0 @@
-export * from './SchemaMigration';
