--- packages/graphback-db-manage/src/providers/index.ts
@@ -1,2 +1,0 @@
-export * from './MigrationProvider';
-export * from './KnexMigrationProvider';
